Plant and Weed objects were detected.
Once we capture the image of the plant in a set area (say 20*20 centimetre square). We can utilise image augmentation techniques, smoothen the image, and identify blobs smaller than the main central plant as "aliens". Hence this is proof of concept that this is a simple implementation of Weed detection.