Update README.md
I was a Research Assistant at BITS Pilani in the Mechanical Engineering Department from August 2022 - December 2022 and worked on the ideation and development of a CNC-Based Automation for Terrace Farming.
YOLOv7 based weed detection was a small part of this project. I created a simple way of detecting weed plants which will grow in and around the main plants, and can be identified by way of image capture --> image augmentation --> and finally use this model to detect the weed and notify the user.